requirements.txt
# Core FastAPI and web dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.4.2
jinja2>=3.1.2
httpx>=0.25.1

# Data processing and scientific computing
numpy>=1.26.2
pandas>=2.1.2
matplotlib>=3.8.1
scikit-learn>=1.3.2

# Core ML and transformers (keeping for Qwen compatibility)
torch>=2.2.0
transformers>=4.37.0
sentence-transformers>=2.3.1
protobuf>=4.25.0

# Quantization support for HuggingFace models
bitsandbytes>=0.42.0
accelerate>=0.25.0
optimum>=1.16.0

# GGUF and llama.cpp support for GPT-OSS 20B
llama-cpp-python>=0.2.0

# Development and monitoring
watchdog>=3.0.0

# Optional: For advanced quantization techniques
# auto-gptq>=0.7.0
# autoawq>=0.1.8